Move out inner class to SystemTestDeployedInstances
